Use double for all drawing methods in Renderer

+++ lib/perobobbot.rendering/src/main/java/perobobbot/rendering/RenderingTools.java
@@ -2,7 +2,7 @@ package perobobbot.rendering;
 
 public class RenderingTools {
 
-    public static double computeCellSize(int fullSize, int nbCells, double margin, double spacing) {
+    public static double computeCellSize(double fullSize, int nbCells, double margin, double spacing) {
         return (fullSize - margin * 2 + spacing) / nbCells - spacing;
     }
 }
